refactor: separtaes the implementation between the alu as an API and the CLI that will have the main program and will also have the interaction with he user.
